Comment Core API (including internal one)
