+++ core/src/common/logger.ts
@@ -1,1 +1,4 @@
+/**
+ * Logs app events to stdout.
+ */
 export const log = require('debug')('mongo-seeding');
