+++ core/src/config.ts
@@ -2,6 +2,9 @@ import * as extend from 'extend';
 import { SeederCollection, DeepPartial } from './common';
 import { SeederDatabaseConfig, defaultDatabaseConfigObject } from './database';
 
+/**
+ * Defines configuration for database seeding.
+ */
 export interface SeederConfig {
   database: SeederDatabaseConfig; // database connection URI or configuration object
   databaseReconnectTimeout: number; // maximum time of waiting for successful MongoDB connection in milliseconds
@@ -9,6 +12,9 @@ export interface SeederConfig {
   dropCollections: boolean; // drops collection before importing it
 }
 
+/**
+ * Stores default configuration for database seeding.
+ */
 export const defaultSeederConfig: SeederConfig = {
   database: defaultDatabaseConfigObject,
   databaseReconnectTimeout: 10000,
@@ -16,6 +22,12 @@ export const defaultSeederConfig: SeederConfig = {
   dropCollections: false,
 };
 
+/**
+ * Merges configuration for database seeding.
+ *
+ * @param partial Partial config object. If not specified, returns a default config object.
+ * @param previous Previous config object. If not specified, uses a default config object as a base.
+ */
 export const mergeSeederConfig = (
   partial?: DeepPartial<SeederConfig>,
   previous?: SeederConfig,
@@ -30,16 +42,28 @@ export const mergeSeederConfig = (
   return extend(true, config, source, partial);
 };
 
+/**
+ * Defines collection reading configuration.
+ */
 export interface SeederCollectionReadingOptions {
   extensions: string[]; // files that should be imported
   transformers: Array<(collection: SeederCollection) => SeederCollection>; // optional transformer functions
 }
 
+/**
+ * Stores default collection reading configuration values.
+ */
 export const defaultCollectionReadingOptions: SeederCollectionReadingOptions = {
   extensions: ['json', 'js'],
   transformers: [],
 };
 
+/**
+ * Merges configuration for collection reading.
+ *
+ * @param partial Partial config object. If not specified, returns a default config object.
+ * @param previous Previous config object. If not specified, uses a default config object as a base.
+ */
 export const mergeCollectionReadingOptions = (
   partial?: DeepPartial<SeederCollectionReadingOptions>,
   previous?: SeederCollectionReadingOptions,
