+++ core/src/database/config.ts
@@ -1,23 +1,75 @@
+/**
+ * Represents database connection configuration. It can be a URI string or object.
+ */
 export type SeederDatabaseConfig = string | SeederDatabaseConfigObject;
 
-export interface SeederDatabaseConfigObjectOptions {
-  [key: string]: string;
-}
-
+/**
+ * Defines configuration for Database connection in a form of an object.
+ */
 export interface SeederDatabaseConfigObject {
+  /**
+   * Database connection protocol
+   */
   protocol: string;
+
+  /**
+   * Database connection host
+   */
   host: string;
+
+  /**
+   * Database connection port
+   */
   port: number;
+
+  /**
+   * Database name.
+   */
   name: string;
+
+  /**
+   * Database Username.
+   */
   username?: string;
+
+  /**
+   * Database password.
+   */
   password?: string;
 
   /**
-   *  See all options for Database Connection URI: https://docs.mongodb.com/manual/reference/connection-string
+   * Options for MongoDB Database Connection URI.
+   * Read more on: https://docs.mongodb.com/manual/reference/connection-string.
    */
   options?: SeederDatabaseConfigObjectOptions;
 }
 
+/**
+ * Defines options for MongoDB Database Connection URI.
+ * Read more on: https://docs.mongodb.com/manual/reference/connection-string.
+ */
+export interface SeederDatabaseConfigObjectOptions {
+  [key: string]: string;
+}
+
+/**
+ * Stores default values for database connection.
+ */
+export const defaultDatabaseConfigObject: SeederDatabaseConfigObject = {
+  protocol: 'mongodb',
+  host: '127.0.0.1',
+  port: 27017,
+  name: 'database',
+  username: undefined,
+  password: undefined,
+  options: undefined,
+};
+
+/**
+ * Checks if an object is valid database connection configuration.
+ *
+ * @param object Input object
+ */
 export function isSeederDatabaseConfigObject(
   object: any,
 ): object is SeederDatabaseConfigObject {
@@ -28,13 +80,3 @@ export function isSeederDatabaseConfigObject(
     typeof object.name === 'string'
   );
 }
-
-export const defaultDatabaseConfigObject: SeederDatabaseConfigObject = {
-  protocol: 'mongodb',
-  host: '127.0.0.1',
-  port: 27017,
-  name: 'database',
-  username: undefined,
-  password: undefined,
-  options: undefined,
-};
