+++ core/src/database/index.ts
@@ -1,4 +1,4 @@
 export * from './config';
 export * from './database';
 export * from './database-connector';
-export * from './time-utils';
+export * from './time';
