--- core/src/database/time-utils.ts
@@ -1,9 +1,0 @@
-export const sleep = (millis: number): Promise<void> =>
-  new Promise(resolve => setTimeout(resolve, millis));
-
-export const checkTimeoutExpired = (
-  startMillis: number,
-  reconnectTimeoutMillis: number,
-) => {
-  return new Date().getTime() - startMillis >= reconnectTimeoutMillis;
-};
