+++ core/src/populator/filesystem.ts
@@ -2,39 +2,73 @@ import { readdirSync, lstatSync, readFileSync } from 'fs';
 import { EJSON } from 'bson';
 import { extname } from 'path';
 
+/**
+ * Provides functionality for manipulating files and directories.
+ */
 export class FileSystem {
   static FILE_NAME_SPLIT_CHARACTER = '.';
 
+  /**
+   * Lists file names for a given path.
+   *
+   * @param path File path
+   */
   listFileNames(path: string) {
     return readdirSync(path) || [];
   }
 
-  listValidDirectories(inputDirectory: string) {
-    const filesAndDirectories = this.listFileNames(inputDirectory);
+  /**
+   * Lists directories which are not empty or hidden.
+   *
+   * @param path File path
+   */
+  listValidDirectories(path: string) {
+    const filesAndDirectories = this.listFileNames(path);
     const directories = filesAndDirectories.filter(fileOrDirectory => {
-      const path = `${inputDirectory}/${fileOrDirectory}`;
+      const itemPath = `${path}/${fileOrDirectory}`;
       return (
-        this.isDirectory(path) &&
-        !this.isEmpty(path) &&
+        this.isDirectory(itemPath) &&
+        !this.isEmpty(itemPath) &&
         !this.isHidden(fileOrDirectory)
       );
     });
     return directories;
   }
 
+  /**
+   * Checks if a directory is empty.
+   *
+   * @param path File path
+   */
   isEmpty(path: string) {
     return this.listFileNames(path).length === 0;
   }
 
+  /**
+   * Checks if a file or directory is empty.
+   *
+   * @param fileOrDirectoryName File or directory name
+   */
   isHidden(fileOrDirectoryName: string) {
     return fileOrDirectoryName.startsWith('.');
   }
 
+  /**
+   * Checks if a path is a directory.
+   *
+   * @param path File path
+   */
   isDirectory(path: string) {
     const stats = lstatSync(path);
     return stats.isDirectory();
   }
 
+  /**
+   * Filters files which are supported for database seeding.
+   *
+   * @param fileNames Array of file names
+   * @param supportedExtensions Supported file extensions array
+   */
   filterSupportedDocumentFileNames(
     fileNames: string[],
     supportedExtensions: string[],
@@ -56,30 +90,48 @@ export class FileSystem {
     return documentFileNames;
   }
 
+  /**
+   * Checks if a file should be ignored based on extension.
+   *
+   * @param fileNameSplitByDot Array, which comes from file name split by dot
+   */
   shouldIgnoreFile(fileNameSplitByDot: string[]) {
     const isHidden = !fileNameSplitByDot[0];
     const hasNoExtension = fileNameSplitByDot.length === 1;
     return isHidden || hasNoExtension;
   }
 
-  readFilesContent(filePaths: string[]) {
-    return filePaths.reduce<any[]>((arr: any[], filePath) => {
-      const fileContent: any = this.readFile(filePath);
+  /**
+   * Reads content for multiple files.
+   *
+   * @param paths Array of paths
+   */
+  readFilesContent(paths: string[]) {
+    return paths.reduce<any[]>((arr: any[], path) => {
+      const fileContent: any = this.readFile(path);
       return arr.concat(fileContent);
     }, []);
   }
 
-  readFile(filePath: string): any {
-    const fileExtension = extname(filePath);
+  /**
+   * Reads a file from path. If it is a JSON, uses EJSON parser.
+   *
+   * @param path File path
+   */
+  readFile(path: string): any {
+    const fileExtension = extname(path);
     if (fileExtension === '.json') {
-      const content = readFileSync(filePath, 'utf-8');
+      const content = readFileSync(path, 'utf-8');
       return EJSON.parse(content, {
         relaxed: true,
       });
     }
 
-    return require(filePath);
+    return require(path);
   }
 }
 
+/**
+ * Default `FileSystem` instance
+ */
 export const fileSystem = new FileSystem();
